requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.2.1
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
eventlet==0.40.1
executing==2.2.0
flake8==7.3.0
gevent==25.5.1
greenlet==3.2.3
idna==3.10
inflection==0.5.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rest-framework-simplejwt==0.0.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.1
zope.interface==7.2
